requirements.txt
Pillow==9.4.0
pyautogui==0.9.54
keyboard==0.13.5